feat(admin): explicit question range mentioned (#32)
The range of 0-10 is not explicit in the current form. Adding this in
the translation files makes this explicit in practice. It would be more
ideal to add a slider to the UI, but that would probably be a more
extensive solution.

Related issue: https://github.com/adfinis-sygroup/freedomvote/issues/16