feat(states): support for multiple states
A politician can now be electable in multiple states, by having a N:N
relationship between politicians and states.

To support this change, the number of states will be shown for the politician if
multiple are defined, rather than the name of the single state. In the detailed
view a popover is available, to get insight in all the states in which the
politician is active.

In the back-end, there is now a list selector available, and multiple states can
be selected by holding down the Ctrl key. In the database this equates to a
linking table between declared politicians and declared states, both referenced
by index.

The sorting mechanism on the candidates view is not affected by this change.

Related issue: https://github.com/freedomvote/freedomvote/issues/18