feat(states): support for multiple states (#38)
* feat(states): support for multiple states

A politician can now be electable in multiple states, by having a N:N
relationship between politicians and states.

To support this change, the number of states will be shown for the politician if
multiple are defined, rather than the name of the single state. In the detailed
view a popover is available, to get insight in all the states in which the
politician is active.

In the back-end, there is now a list selector available, and multiple states can
be selected by holding down the Ctrl key. In the database this equates to a
linking table between declared politicians and declared states, both referenced
by index.

The sorting mechanism on the candidates view is not affected by this change.

Related issue: https://github.com/freedomvote/freedomvote/issues/18

* fix(states): translation and double count

Based on the review by @winged, prevent the double count operation, and enable
better translations regarding numbers. Just cranking up the code quality.

* chore(states): fixme and compiles messages

- FIXME added to notify about the comment by @winged on
  https://github.com/freedomvote/freedomvote/pull/38 Better practice is to
  include the decimal inside the translation to enable different placement in
  the translation.
- Messages compiled, which results in updated dates and line numbers.