wip(meta): got django-meta package to work
Achieved a basic django-meta flow through views, for which the meta-data is
rendered via an indirection in the index.html base template. Default properties
can be set in the settings.py, although these are only rendered if a page
includes some meta context.
This WIP implementation only renders for /candidates