feat(meta): default meta property in the views
This commit provides a default_meta metadata property for the public views,
which is rendered by default. In this way the integration with social media
platforms is improved. In the future this feature can be expanded to become more
dynamic based on the delivered content.

The default slogan is up for translation.